+++ custom_components/zeekr_ev/const.py
@@ -23,9 +23,11 @@ LOCK = "lock"
 CLIMATE = "climate"
 SWITCH = "switch"
 COVER = "cover"
+DATETIME = "datetime"
 NUMBER = "number"
 SELECT = "select"
-PLATFORMS = [BINARY_SENSOR, BUTTON, CLIMATE, COVER, DEVICE_TRACKER, LOCK, NUMBER, SELECT, SENSOR, SWITCH]
+TIME = "time"
+PLATFORMS = [BINARY_SENSOR, BUTTON, CLIMATE, COVER, DATETIME, DEVICE_TRACKER, LOCK, NUMBER, SELECT, SENSOR, SWITCH, TIME]
 
 
 # Configuration and options
